// frontend/atcoder-editorial-problems/src/generated/orval/models/problem.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * atcoder-editorial-problems-api
 * A definition of AtCoder Editorial Problems API.
 * OpenAPI spec version: 1.0
 */

export interface Problem {
  /** contest category */
  category: string;
  /** contest name */
  contest: string;
  /** problem difficulty */
  difficulty: number;
  /** editorial type */
  editorial_types: string[];
  /** problem id with a editorial writer */
  id: string;
  /** whether this problem is experimental */
  is_experimental: boolean;
  /** problem name */
  name: string;
  /** problem id */
  problem_id: string;
  /** problem index in the contest */
  problem_index: string;
  /** result of user submissions */
  result_code?: string;
  /** contest start time (UNIX time) */
  start_epoch_second: number;
}